fix: treat empty node catalog as inconclusive for trigger filters
Avoid skipping eager paging and flashing "No run matched" while canvas
nodes are still loading by passing nodeCatalogSize into trigger matchability.
